PR-536602: updating packages and fixing vendors.css error.
