+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aleleba/create-react-ssr",
-  "version": "3.1.0",
+  "version": "3.1.1",
   "description": "Starter Kit of server side render of react",
   "bin": "./bin/cli.js",  
   "main": "src/server/index",
@@ -68,7 +68,7 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^28.1.6",
-    "@types/node": "^18.6.5",
+    "@types/node": "^18.7.0",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@types/webpack": "^5.28.0",
