--- src/server/getHashManifest.ts
@@ -1,13 +1,0 @@
-import fs from 'fs';
-
-const getHashManifest = () => {
-	try {
-		const baseUrl = __dirname.replace(/\/server(.*)/,'');
-		const fullURL = `${baseUrl}/assets/manifest-hash.json` ;
-		return JSON.parse(fs.readFileSync(fullURL).toString());
-	}catch(err){
-		console.error(err);
-	}
-};
-
-export default getHashManifest;
